refactor: move createBasicGate and createCompositeGate from RenderPoint.js to gates.js (closes #53)
