+++ logic/gates.js
@@ -1,5 +1,5 @@
 import { Wire } from "./wire.js";
-import { evaluateAll, settleCircuit } from "./evaluate.js";
+import { evaluateAll } from "./evaluate.js";
 
 class Logic {
   static nextId = 1;
@@ -204,4 +204,36 @@ export class CompositeGate extends ConnectableGate {
 
     return this.tempOutput;
   }
+}
+
+
+/**
+ * Creates a basic gate instance based on the given type.
+ * 
+ * @param {string} type - The type of gate ("input", "output", "clock" or logic gate type)
+ * @returns {Input|Output|Clock|Gate} The instantiated gate object
+ */
+export function createBasicGate(type) {
+  return type === "input"
+    ? new Input(false)
+    : type === "output"
+    ? new Output()
+    : type === "clock"
+    ? new Clock(false)
+    : new Gate(type, []);
+}
+
+/**
+ * Creates a composite gate from saved circuit data. 
+ * 
+ * @param {string} name - The label/name of the composite gate
+ * @param {Object} circuitData - The circuit data used to construct the composite gate
+ * @returns {CompositeGate} The instantiated composite gate object 
+ */
+export function createCompositeGate(name, circuitData) {
+  const inputs = new Array(circuitData.renderNodes.filter(n => n.gate.type === "input").length);
+  const gate = new CompositeGate(inputs, circuitData);
+  gate.label = name;
+  gate.parseCircuitData();
+  return gate;
 }
